Add ability to only pull new articles since the last time checked. Resolves #3.
The change was relatively simple:
  - Add a new field to the handler
  - In main.rs load the handler earlier
  - Collect the since value from the response to the retrieve request
  - Pass that value to the handler

Additionally, working on this exposed an issue in the config saving
method, and the associated test! The test was not actually checking that
a written configuration was properly loaded. The problem was in the way
the file was written.

+++ rePocket/src/pocketquery/mod.rs
@@ -71,7 +71,7 @@ pub struct PocketQuery {
     #[serde(skip_serializing_if = "Option::is_none")]
     domain: Option<String>,
     // Although Pocket expects an integer, perhaps we can make it nice, like a DateTime<Local>
-    since: Option<usize>,
+    since: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     count: Option<u8>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -91,7 +91,7 @@ impl PocketQuery {
         detail_type: Option<QDetailType>,
         search: Option<String>,
         domain: Option<String>,
-        since: Option<usize>,
+        since: Option<u64>,
         count: Option<u8>,
         offset: Option<u32>,
         total: Option<QBool>,
@@ -126,7 +126,7 @@ pub struct QueryBuilder {
     detail_type: Option<QDetailType>,
     search: Option<String>,
     domain: Option<String>,
-    since: Option<usize>,
+    since: Option<u64>,
     count: Option<u8>,
     offset: Option<u32>,
     total: Option<QBool>,
@@ -185,7 +185,7 @@ impl QueryBuilder {
     }
 
     // Perhaps timestamp should be of a type similar to whatever the crate chrono uses.
-    pub fn set_since(mut self, timestamp: usize) -> Self {
+    pub fn set_since(mut self, timestamp: u64) -> Self {
         self.since = Some(timestamp);
 
         self
